add JMH test
