+++ common/src/test/java/io/netty/util/NormalizeTicksPerWheelBenchmark.java
@@ -1,0 +1,62 @@
+package io.netty.util;
+
+import org.openjdk.jmh.annotations.*;
+import org.openjdk.jmh.runner.Runner;
+import org.openjdk.jmh.runner.RunnerException;
+import org.openjdk.jmh.runner.options.Options;
+import org.openjdk.jmh.runner.options.OptionsBuilder;
+
+import java.util.concurrent.TimeUnit;
+
+
+@BenchmarkMode(Mode.Throughput)
+@Warmup(iterations = 3)
+@Measurement(iterations = 10, time = 5, timeUnit = TimeUnit.SECONDS)
+@Threads(8)
+@Fork(2)
+@OutputTimeUnit(TimeUnit.MILLISECONDS)
+public class NormalizeTicksPerWheelBenchmark {
+
+    @Benchmark
+    public void testNew() {
+        int count = 10000000;
+        while (count > 0) {
+            oldNormalizeTicksPerWheel(count);
+        }
+    }
+
+    @Benchmark
+    public void testOld() {
+        int count = 10000000;
+        while (count > 0) {
+            count--;
+            oldNormalizeTicksPerWheel(count);
+        }
+    }
+
+    private static int oldNormalizeTicksPerWheel(int ticksPerWheel) {
+        int normalizedTicksPerWheel = 1;
+        while (normalizedTicksPerWheel < ticksPerWheel) {
+            normalizedTicksPerWheel <<= 1;
+        }
+        return normalizedTicksPerWheel;
+    }
+
+    private static int newNormalizeTicksPerWheel(int ticksPerWheel) {
+        int normalizedTicksPerWheel = ticksPerWheel - 1;
+        normalizedTicksPerWheel |= normalizedTicksPerWheel >>> 1;
+        normalizedTicksPerWheel |= normalizedTicksPerWheel >>> 2;
+        normalizedTicksPerWheel |= normalizedTicksPerWheel >>> 4;
+        normalizedTicksPerWheel |= normalizedTicksPerWheel >>> 8;
+        normalizedTicksPerWheel |= normalizedTicksPerWheel >>> 16;
+        return normalizedTicksPerWheel + 1;
+    }
+
+    public static void main(String[] args) throws RunnerException {
+        Options opt = new OptionsBuilder()
+                .include(NormalizeTicksPerWheelBenchmark.class.getSimpleName())
+                .build();
+
+        new Runner(opt).run();
+    }
+}
